% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R
\name{blend}
\alias{blend}
\title{blend}
\usage{
blend(
  name = "fig",
  folder_out = "output",
  folder_animate = NULL,
  fps = 2,
  morph = F,
  frames = 10,
  folder1 = NULL,
  colors1 = NULL,
  folder2 = NULL,
  colors2 = NULL,
  folder3 = NULL,
  colors3 = NULL,
  folder4 = NULL,
  colors4 = NULL,
  folder5 = NULL,
  colors5 = NULL
)
}
\arguments{
\item{name}{Default = "fig". Base name of figure. Will be appended with numbers for each blend.}

\item{folder_out}{Default = "output". Name of folder to save figs in.}

\item{folder_animate}{Default = NULL. Path to folder with animations}

\item{fps}{Default = 2. Frames per second for animations}

\item{morph}{Default = F. Morph animations}

\item{frames}{Default = 10. Frames used for morphing}

\item{folder1}{Default = NULL. Folder with images for layer 1.}

\item{colors1}{Default = NULL. (OPTIONAL) List of colors to combine for layer 1 images.}

\item{folder2}{Default = NULL. Folder with images for layer 2.}

\item{colors2}{Default = NULL. (OPTIONAL) List of colors to combine for layer 2 images.}

\item{folder3}{Default = NULL. Folder with images for layer 3.}

\item{colors3}{Default = NULL. (OPTIONAL) List of colors to combine for layer 3 images.}

\item{folder4}{Default = NULL. Folder with images for layer 4.}

\item{colors4}{Default = NULL. (OPTIONAL) List of colors to combine for layer 4 images.}

\item{folder5}{Default = NULL. Folder with images for layer 5.}

\item{colors5}{Default = NULL. (OPTIONAL) List of colors to combine for layer 5 images.}
}
\value{
blended images
}
\description{
Blend multiple layers of images into various combinations
}
\keyword{blend,}
\keyword{merge,}
\keyword{nft}
