% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardragonox.R
\docType{package}
\name{cardragonox}
\alias{cardragonox}
\title{cardragonox}
\description{
\itemize{
\item Github: https://github.com/cardragonox/cardragonox
\item Webpage: https://cardragonox.github.io/
}
}
